ft(docker): setup an nginx proxy for the spotlights app with docker
- configure default nginx config file
- build both nodejs and nginx docker images

+++ deploy_script.sh
@@ -4,17 +4,25 @@
 set -e
 
 # Where to deploy our site on our server
-ASSETS_SOURCE_URL="http://178.128.102.213/"
 cd ../build.zip
-unzip build.zip
+unzip -o build.zip
 cd build || exit
 
 # Build the docker image
-docker stop datahub-spotlights
-docker rm datahub-spotlights
-docker rmi datahub-spotlights-image
-docker build -t datahub-spotlights-image .
-docker run -d -p 3000:3000 --name datahub-spotlights -e ASSETS_SOURCE_URL="$ASSETS_SOURCE_URL" datahub-spotlights-image:latest
+if [ ! "$(docker ps -q -f name=datahub-spotlights)" ]; then
+    if [ "$(docker ps -aq -f status=exited -f name=datahub-spotlights)" ]; then
+        # cleanup
+        docker stop datahub-spotlights
+        docker rm datahub-spotlights
+        docker rmi datahub-spotlights-image
+    fi
+    # run container
+    docker-compose build
+    docker-compose up -d
+fi
+#docker run -d -p 3000:3000 --name datahub-spotlights -e ASSETS_SOURCE_URL="$ASSETS_SOURCE_URL" datahub-spotlights-image:latest
 # clean up scripts
+
+
 cd ~
 rm -rf deploy_spotlights
